+++ CashBox/DatabaseProject/Models/DailyClose.cs
@@ -4,8 +4,8 @@ namespace DatabaseProject.Models
 {
     public class DailyClose : BaseEntity
     {
-        public double InitialAmount { get; set; }
-        public double? FinalAmount { get; set; }
+        public decimal InitialAmount { get; set; }
+        public decimal? FinalAmount { get; set; }
         public long CasherId { get; set; }
 
         [ForeignKey(nameof(CasherId))]
